first version of navbar desktop
