+++ apps/app-web/components/common/NavBar/desktop.module.css
@@ -1,0 +1,72 @@
+.navigation {
+    width: 100%;
+    height: 4rem ;
+    background-color: var(--color-background);
+    border-bottom: 1px solid #E0E0E0;
+    box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.5);
+}
+.navigationContainer{
+    position: relative;
+    display: flex;
+    width: 90%;
+    height: 100%;
+    align-items: center;
+    justify-content: space-between;
+    margin: 0 auto;
+}
+.sideContainer {
+    display: flex;
+    flex: 1;
+}
+.sideContainer:last-of-type {
+    justify-content: flex-end;
+}
+.navigation ul {
+    margin: 0;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 2rem;
+}
+.navigation ul li {
+    list-style: none;
+    padding: 0 0.5rem;
+}
+.navigation ul li a {
+    text-align: center;
+    text-decoration: none;
+    color: var(--text-secondary);
+    user-select: none;
+    font-size: 1rem;
+    line-height: 3.75rem;
+    font-weight: 400;
+    letter-spacing: 0.05rem;
+    text-align: center;
+    user-select: none;
+}
+/*to review the hover color*/
+.navigation ul li a:hover {
+    color: var(--color-primary);
+    opacity: 0.8;
+}
+.navigation ul li.active a {
+    color: var(--color-primary);
+}
+.indicator {
+    position: absolute;
+    top: 70%;
+    width: 60px;
+    border: 2px solid var(--color-primary);
+    background-color: var(--color-background);
+    transition: 0.5s;
+}
+.navigation ul li:nth-child(1).active~.indicator {
+    transform: translateX(calc(-26px * 4));
+}
+.navigation ul li:nth-child(2).active~.indicator {
+    transform: translateX(calc(-4px * 1));
+}
+.navigation ul li:nth-child(3).active~.indicator {
+    transform: translateX(calc(-25px * -4));
+}
