+++ apps/app-web/pages/_app.tsx
@@ -1,8 +1,7 @@
 import { SessionProvider } from "next-auth/react";
 import type { AppProps } from "next/app";
 import "../styles/globals.css";
-import NavBar from "../components/common/NavBar";
-
+import NavBar from "../components/common/NavBar/NavBar";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
