+++ apps/app-web/styles/globals.css
@@ -17,112 +17,3 @@ body {
 .main-content{
     padding-bottom: 80px;
 }
-
-.navigation {
-    position: relative;
-    width: 100%;
-    height: 4rem ;
-    border-top-right-radius: 10px;
-    border-top-left-radius: 10px;
-    background-color: var(--color-background);
-    border: 1px solid #E0E0E0;
-    box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.5);
-}
-.navigation-mobile {
-    position: fixed;
-    bottom: 0;
-}
-.navigation ul {
-    width: 100%;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-.navigation ul li {
-    position: relative;
-    list-style: none;
-    min-width: 48px;
-    min-height: 48px;;
-    width: 3rem;
-    height: 3rem;
-    z-index: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0 0.5rem;
-}
-.navigation ul li a {
-    position: relative;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    width: 100%;
-    text-align: center;
-    text-decoration: none;
-    color: #fff;
-    user-select: none;
-}
-.navigation ul li a .icon {
-    position: relative;
-    display: block;
-    line-height: 75px;
-    font-size: 1.5rem;
-    text-align: center;
-    color: var(--text-primary);
-    font-weight: 700;
-    text-decoration: none;
-    transition: all 0.5s ease; 
-    user-select: none;
-}
-
-.navigation ul li a .label{
-    position: absolute;
-    color: var(--text-secondary);
-    font-size: 0.625rem;
-    line-height: 1.25;
-    font-weight: 400;
-    letter-spacing: 0.05rem;
-    transition: 0.5s;
-    opacity: 0.3;
-    transform: translateY(20px);
-    text-align: center;
-    user-select: none;
-}
-.navigation ul li.active a .icon {
-    transform: translateY(-20px);
-    color: var(--color-secondary);
-}
-.navigation ul li.active a .label {
-    opacity: 1;
-    transform: translateY(0px);
-    color: var(--text-primary);    
-}
-.indicator {
-    position: absolute;
-    top: -50%;
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    border: 6px solid var(--color-secondary);
-    background-color: var(--color-background);
-    transition: 0.5s;
-}
-.navigation ul li:nth-child(1).active~.indicator {
-    transform: translateX(calc(-32px * 4));
-}
-.navigation ul li:nth-child(2).active~.indicator {
-    transform: translateX(calc(-32px * 2));
-}
-.navigation ul li:nth-child(3).active~.indicator {
-    transform: translateX(calc(-35px * 0));
-}
-.navigation ul li:nth-child(4).active~.indicator {
-    transform: translateX(calc(-32px * -2));
-}
-.navigation ul li:nth-child(5).active~.indicator {
-    transform: translateX(calc(-32px * -4));
-}
