+++ apps/app-web/styles/tokens.css
@@ -1,8 +1,8 @@
 :root {
   /* Colors */
   --color-background: #FAF9F6;
-  --color-primary:    #A8B39A;
-  --color-secondary:  #D75A4A;
+  --color-primary:    #D75A4A;
+  --color-secondary:  #A8B39A;
   --text-primary:     #333333;
   --text-secondary:   #666666;
